requirements.txt
aboutcode.pipeline==0.1.0
aboutcode-toolkit==10.1.0
alabaster==0.7.13
anyio==4.1.0
asgiref==3.7.2
attrs==23.1.0
Babel==2.14.0
boolean.py==4.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.1
dateparser==1.1.8
Deprecated==1.2.14
Django==5.0.1
django-environ==0.11.2
django-oauth-toolkit==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
django-ninja==1.2.1
docutils==0.20.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
execnet==2.0.2
gitdb==4.0.10
GitPython==3.1.40
gunicorn==21.2.0
h11==0.14.0
http-message-signatures==0.4.4
http-sfv==0.9.8
httpcore==0.17.3
httpx==0.25.1
idna==3.4
imagesize==1.4.1
importlib-metadata==7.0.1
iniconfig==2.0.0
isort==5.12.0
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
jwcrypto==1.5.0
keyring==24.3.0
license-expression==30.2.0
livereload==2.6.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==10.1.0
mypy-extensions==1.0.0
nh3==0.2.15
oauthlib==3.2.2
openpyxl==3.1.2
packageurl-python==0.15.6
packaging==23.1
pathspec==0.11.2
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.0.0
psycopg==3.1.16
psycopg-binary==3.1.16
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.17.2
python-dateutil==2.8.2
pydantic==2.8.2
pytz==2023.3
PyYAML==6.0.1
readme-renderer==42.0
regex==2023.8.8
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
saneyaml==0.6.0
SecretStorage==3.3.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.4
twine==4.0.2
typing_extensions==4.6.3
tzlocal==5.0.1
unidiff==0.7.5
urllib3==2.0.3
wrapt==1.15.0
zipp==3.17.0
